refactor(session): extract last message snippet module
Move lastMessageSnippet hydration, recent-message scanning, and snippet tests out of session_manager.rs into a focused session module while keeping list-session behavior unchanged.
